+++ components/ui/molecules/index.ts
@@ -1,3 +1,4 @@
 export { default as TabBar } from './TabBar'
 export { default as TabItem } from './TabItem'
 export { default as Tag } from './Tag'
+export { default as NavBar } from './NavBar'
